js module, arreglar algunas funciones
