--- src/assets/my-icons-collection (1)/backup.txt
@@ -1,1 +1,0 @@
-eyIyIjp7IklEIjoyLCJuYW1lIjoiTXkgaWNvbnMgY29sbGVjdGlvbiIsImJvb2ttYXJrX2lkIjo2MzQ5OTMyLCJjcmVhdGVkIjoxNTU2MDc0NDY0LCJ1cGRhdGVkIjoxNTU2MDc0NDY4LCJhY3RpdmUiOjEsInNvdXJjZSI6ImxvY2FsIiwib3JkZXIiOjAsImNvbG9yIjoiMDAwMDAwIiwic3RhdHVzIjoxfSwiNjM0OTkzMiI6W3siaWQiOjE0NTgwMiwidGVhbSI6MCwibmFtZSI6ImZhY2Vib29rLWNvbG9yIiwiY29sb3IiOiIjWFhYWFhYIiwicHJlbWl1bSI6MCwic29ydCI6MX0seyJpZCI6Njk0MDcsInRlYW0iOjAsIm5hbWUiOiJmYWNlYm9vay1idyIsImNvbG9yIjoiIzAwMDAwMCIsInByZW1pdW0iOjAsInNvcnQiOjJ9LHsiaWQiOjYwNTQzLCJ0ZWFtIjowLCJuYW1lIjoiZ21haWwtYnciLCJjb2xvciI6IiMwMDAwMDAiLCJwcmVtaXVtIjowLCJzb3J0Ijo0fSx7ImlkIjo4ODg4NTMsInRlYW0iOjAsIm5hbWUiOiJnbWFpbC1jb2xvciIsImNvbG9yIjoiI1hYWFhYWCIsInByZW1pdW0iOjAsInNvcnQiOjN9LHsiaWQiOjE4MTU0MSwidGVhbSI6MCwibmFtZSI6ImltYWdlLWJ3IiwiY29sb3IiOiIjMDAwMDAwIiwicHJlbWl1bSI6MCwic29ydCI6NX0seyJpZCI6MTM4NTcwLCJ0ZWFtIjowLCJuYW1lIjoiaW1hZ2UtY29sb3IiLCJjb2xvciI6IiNYWFhYWFgiLCJwcmVtaXVtIjowLCJzb3J0Ijo2fSx7ImlkIjo0NzAzNDAsInRlYW0iOjAsIm5hbWUiOiJsaWtlLWNvbG9yIiwiY29sb3IiOiIjWFhYWFhYIiwicHJlbWl1bSI6MCwic29ydCI6N30seyJpZCI6MTg5NjcxLCJ0ZWFtIjowLCJuYW1lIjoiaGVhcnQiLCJjb2xvciI6IiNYWFhYWFgiLCJwcmVtaXVtIjowLCJzb3J0Ijo4fSx7ImlkIjoxMTU5NjMzLCJ0ZWFtIjowLCJuYW1lIjoiZWRpdC1idyIsImNvbG9yIjoiIzAwMDAwMCIsInByZW1pdW0iOjAsInNvcnQiOjl9LHsiaWQiOjEwNDAyMjgsInRlYW0iOjAsIm5hbWUiOiJlZGl0LWNvbG9yIiwiY29sb3IiOiIjWFhYWFhYIiwicHJlbWl1bSI6MCwic29ydCI6MTB9LHsiaWQiOjU2ODUyMywidGVhbSI6MCwibmFtZSI6ImRpc2xpa2UiLCJjb2xvciI6IiNYWFhYWFgiLCJwcmVtaXVtIjowLCJzb3J0IjoxMX0seyJpZCI6MTc0ODU1LCJ0ZWFtIjowLCJuYW1lIjoiaW5zdGFncmFtLWNvbG9yIiwiY29sb3IiOiIjWFhYWFhYIiwicHJlbWl1bSI6MCwic29ydCI6MTJ9LHsiaWQiOjEzODQwMTUsInRlYW0iOjAsIm5hbWUiOiJpbnN0YWdyYW0tYnciLCJjb2xvciI6IiMwMDAwMDAiLCJwcmVtaXVtIjowLCJzb3J0IjoxM30seyJpZCI6Njk0ODAsInRlYW0iOjAsIm5hbWUiOiJ0d2l0dGVyLWJ3IiwiY29sb3IiOiIjMDAwMDAwIiwicHJlbWl1bSI6MCwic29ydCI6MTR9LHsiaWQiOjE0NTgxMiwidGVhbSI6MCwibmFtZSI6InR3aXR0ZXItY29sb3IiLCJjb2xvciI6IiNYWFhYWFgiLCJwcmVtaXVtIjowLCJzb3J0IjoxNX1dfQ==
