--- src/assets/my-icons-collection (1)/font/_flaticon.scss
@@ -1,52 +1,0 @@
-    /*
-    Flaticon icon font: Flaticon
-    Creation date: 24/04/2019 02:54
-    */
-
-    @font-face {
-  font-family: "Flaticon";
-  src: url("./Flaticon.eot");
-  src: url("./Flaticon.eot?#iefix") format("embedded-opentype"),
-       url("./Flaticon.woff2") format("woff2"),
-       url("./Flaticon.woff") format("woff"),
-       url("./Flaticon.ttf") format("truetype"),
-       url("./Flaticon.svg#Flaticon") format("svg");
-  font-weight: normal;
-  font-style: normal;
-}
-
-@media screen and (-webkit-min-device-pixel-ratio:0) {
-  @font-face {
-    font-family: "Flaticon";
-    src: url("./Flaticon.svg#Flaticon") format("svg");
-  }
-}
-
-    .fi:before{
-        display: inline-block;
-  font-family: "Flaticon";
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  line-height: 1;
-  text-decoration: inherit;
-  text-rendering: optimizeLegibility;
-  text-transform: none;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  font-smoothing: antialiased;
-    }
-
-    .flaticon-facebook-bw:before { content: "\f100"; }
-.flaticon-gmail-bw:before { content: "\f101"; }
-.flaticon-image-bw:before { content: "\f102"; }
-.flaticon-edit-bw:before { content: "\f103"; }
-.flaticon-instagram-bw:before { content: "\f104"; }
-.flaticon-twitter-bw:before { content: "\f105"; }
-    
-    $font-Flaticon-facebook-bw: "\f100";
-    $font-Flaticon-gmail-bw: "\f101";
-    $font-Flaticon-image-bw: "\f102";
-    $font-Flaticon-edit-bw: "\f103";
-    $font-Flaticon-instagram-bw: "\f104";
-    $font-Flaticon-twitter-bw: "\f105";
