--- src/assets/my-icons-collection (1)/font/flaticon.css
@@ -1,38 +1,0 @@
-	/*
-  	Flaticon icon font: Flaticon
-  	Creation date: 24/04/2019 02:54
-  	*/
-
-@font-face {
-  font-family: "Flaticon";
-  src: url("./Flaticon.eot");
-  src: url("./Flaticon.eot?#iefix") format("embedded-opentype"),
-       url("./Flaticon.woff2") format("woff2"),
-       url("./Flaticon.woff") format("woff"),
-       url("./Flaticon.ttf") format("truetype"),
-       url("./Flaticon.svg#Flaticon") format("svg");
-  font-weight: normal;
-  font-style: normal;
-}
-
-@media screen and (-webkit-min-device-pixel-ratio:0) {
-  @font-face {
-    font-family: "Flaticon";
-    src: url("./Flaticon.svg#Flaticon") format("svg");
-  }
-}
-
-[class^="flaticon-"]:before, [class*=" flaticon-"]:before,
-[class^="flaticon-"]:after, [class*=" flaticon-"]:after {   
-  font-family: Flaticon;
-        font-size: 20px;
-font-style: normal;
-margin-left: 20px;
-}
-
-.flaticon-facebook-bw:before { content: "\f100"; }
-.flaticon-gmail-bw:before { content: "\f101"; }
-.flaticon-image-bw:before { content: "\f102"; }
-.flaticon-edit-bw:before { content: "\f103"; }
-.flaticon-instagram-bw:before { content: "\f104"; }
-.flaticon-twitter-bw:before { content: "\f105"; }
